refactor(settings): move dns resolution mode to network card
Users look for DNS next to proxy and NAT, not inside the Downloads
engine-tuning form. The control now lives in the Network dialog as its
own DNS section (keys under settings.network.dns.*); the submitted
patch shape ({ engine: { dnsMode } }) and all backend behavior are
unchanged.
